requirements.txt
transformers>=4.36.0
torch>=2.0.0
accelerate>=0.25.0
huggingface-hub>=0.19.0
python-dotenv>=1.0.0 